requirements.txt
python-dotenv
streamlit 
PyPDF2 
google-generativeai